Fixed accidental global assignment loop bug
Previous commit changed how the assignment loop works to fix __newindex metatable behavior.
However, I've noticed it has inadvertently broken other logic because I forgot to include context-awareness to my change. So this commit adds a check for if RHS is greater than 1 to condition when to use Left-To-Right evaluation and when to use Right-To-Left evaluation.

in cases such as:

local a
print(a == nil, a ~= nil)

MoonSharp would incorrectly output "false true" instead of the expected "true false" as a result of the adjusted assignment loop not being properly conditioned before using Left-To-Right evaluation.